Fix 64bit compilation and all new compiler warnings.
git-svn-id: svn://mplayerosx.ch/mplayerosx/trunk@136 6577e75b-6453-0410-a6fa-8f9c828ad7c1

+++ Sources/Classes/FullscreenControls.m
@@ -28,7 +28,7 @@
 
 @implementation FullscreenControls
 
-- (id)initWithContentRect:(NSRect)contentRect styleMask:(unsigned int)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag {
+- (id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag {
 	
     // Make the window borderless
     NSWindow* result = [super initWithContentRect:contentRect styleMask:NSBorderlessWindowMask backing:NSBackingStoreBuffered defer:NO];
